[Bronze II] Title: 이상한 곱셈, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Bronze/1225. 이상한 곱셈/이상한 곱셈.cc
@@ -6,7 +6,7 @@ int main()
     std::string str1, str2;
     std::cin >> str1 >> str2;
     
-    long long sum1, sum2;
+    long sum1, sum2;
     sum1 = sum2 = 0;
     for (int i = 0; i < str1.length(); i++)
         sum1 += (str1[i] - '0');
